requirements.txt
langchain
PyPDF2
streamlit
python-dotenv
cohere

-e .
